requirements.txt
torch
transformers>=4.52.0
Pillow
langchain
imagehash
pymongo[srv]>=4.6.0
streamlit
torchvision
accelerate
modelscope
qwen-vl-utils[decord]==0.0.8
dnspython
tenacity
certifi